chore: pull base setup into separate file
