+++ templates/base/vitest.config.ts
@@ -3,6 +3,7 @@ import { defineWorkersConfig } from "@cloudflare/vitest-pool-workers/config";
 export default defineWorkersConfig(async () => {
   return {
     test: {
+      setupFiles: ["./tests/setup.ts"],
       globals: true,
       poolOptions: {
         workers: {
